+++ ipacm/src/IPACM_Wlan.cpp
@@ -2203,7 +2203,7 @@ int IPACM_Wlan::handle_wlan_client_route_rule_ext_v2(uint8_t *mac_addr, ipa_ip_t
 			return IPACM_FAILURE;
 		}
 
-		rt_rule->rules = (uint64_t)calloc(NUM, sizeof(struct ipa_rt_rule_add_ext_v2));
+		rt_rule->rules = (uintptr_t)calloc(NUM, sizeof(struct ipa_rt_rule_add_ext_v2));
 		if (!rt_rule->rules) {
 			IPACMERR("Failed to allocate memory.\n");
 			free(rt_rule);
@@ -3578,7 +3578,7 @@ int IPACM_Wlan::install_uplink_filter_rule_per_client_v2
 		goto fail;
 	}
 	memset(pFilteringTable, 0, len);
-	pFilteringTable->rules = (uint64_t)calloc(prop->num_ext_props, sizeof(struct ipa_flt_rule_add_v2));
+	pFilteringTable->rules = (uintptr_t)calloc(prop->num_ext_props, sizeof(struct ipa_flt_rule_add_v2));
 	if (!pFilteringTable->rules) {
 		IPACMERR("Failed to allocate memory for filtering rules\n");
 		ret = IPACM_FAILURE;
